Update name of package for GCloud CLI utility
Google has "rebranded" the Google Cloud SDK package and added a new
Homebrew package (`gcloud-cli`). The old one (`google-cloud-sdk`) was
updated to declare the new package as a dependency, but this causes a
persistent `Installing google-cloud-sdk` package when installed from a
Brewfile. Uninstalling the old package and instaslling the new one via
the Brewfile solves that problem.